#7.4 Protecting Video Routes

+++ controllers/videoController.js
@@ -76,7 +76,11 @@ export const getEditVideo = async (req, res) => {
 
     try {
         const video = await Video.findById(id)
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });;
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });;
+        }
     } catch (err) {
         res.redirect(routes.home)
     }
@@ -100,7 +104,12 @@ export const deleteVideoController = async (req, res) => {
         params: { id }
     } = req;
     try {
-        await Video.findOneAndRemove({ _id: id });
+        const video = await Video.findById(id)
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
     } catch (err) {
         console.log(err);
     }
